relations and added authorization for methods

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import { TaskRepository } from './task.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
 import { GetTasksFilterDto } from './dto/get-tasks-filter.dto';
+import { User } from '../auth/entity/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -37,12 +38,18 @@ export class TasksService {
   //     return tasks;
   // }
 
-  async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-    return this.taskRepository.getTasks(filterDto);
+  async getTasks(
+    filterDto: GetTasksFilterDto,
+    user: User,
+  ): Promise<Task[]> {
+    return this.taskRepository.getTasks(filterDto, user);
   }
 
-  async getTaskById(@Param('id', ParseIntPipe) id: number): Promise<Task> {
-    const found = await this.taskRepository.findOne(id);
+  async getTaskById(
+    @Param('id', ParseIntPipe) id: number,
+    user: User,
+  ): Promise<Task> {
+    const found = await this.taskRepository.findOne({ where: { id, userId: user.id } });
 
     if (!found) {
       throw new NotFoundException(`Task ${id} not found`);
@@ -51,48 +58,30 @@ export class TasksService {
     return found;
   }
 
-  async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-    return this.taskRepository.createTask(createTaskDto);
+  async createTask(
+    createTaskDto: CreateTaskDto,
+    user: User,
+  ): Promise<Task> {
+    return this.taskRepository.createTask(createTaskDto, user);
   }
 
-  async deleteTask(@Param('id', ParseIntPipe) id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTask(
+    @Param('id', ParseIntPipe) id: number,
+    user : User
+  ): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId:user.id });
 
     if (result.affected === 0) {
       throw new NotFoundException(`Id number ${id} not found`);
     }
   }
 
-  async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id);
+  async updateTaskStatus(id: number, status: TaskStatus, user: User): Promise<Task> {
+    const task = await this.getTaskById(id,user);
     task.status = status;
     await task.save();
 
     return task;
   }
-
-  // createTask(createTaskDto: CreateTaskDto): Task {
-  //   const { title, description } = createTaskDto;
-  //   const task = {
-  //     id: '1234',
-  //     title,
-  //     description,
-  //     status: TaskStatus.OPEN,
-  //   };
-  //
-  //   this.tasks.push(task)
-  //   return task;
-  //
-  // }
-  //
-  // deleteTask(id:string): void {
-  //   this.tasks = this.tasks.filter(task => { task.id === id; })
-  // }
-  //
-  // updateTaskStatus(id:string, status: TaskStatus): Task {
-  //   const task = this.getTaskById(id);
-  //   task.status = status;
-  //   return task;
-  // }
 
 }
